+++ .github/workflows/ci.yml
@@ -21,7 +21,7 @@ jobs:
       - name: 'Install Linux dependencies'
         if: ${{ runner.os == 'Linux' }}
         run: sudo apt install attr e2fsprogs libfuse-dev
-      - name: 'Install MacOS dependencies'
+      - name: 'Install macOS dependencies'
         if: ${{ runner.os == 'macOS' }}
         run: brew install autoconf automake libtool
       - name: 'Checkout the repository'
