+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "crates/libtsk-sys/sleuthkit"]
+[submodule "vendor/sleuthkit"]
 	path = vendor/sleuthkit
 	url = https://github.com/sleuthkit/sleuthkit
