+++ crates/rrg/Cargo.toml
@@ -67,6 +67,10 @@ path = "../ospect"
 [dependencies.rrg-proto]
 path = "../rrg-proto"
 
+[dependencies.tsk]
+path = "../tsk"
+optional = true
+
 [dependencies.winreg]
 path = "../winreg"
 
@@ -118,10 +122,6 @@ optional = true
 
 [dependencies.sha2]
 version = "0.10.8"
-optional = true
-
-[dependencies.tsk]
-path = "../tsk"
 optional = true
 
 # TODO(https://github.com/google/rrg/issues/47): This should be a dev dependency
